fix hive plugin return data incluing \t
