+++ linkis-engineconn-plugins/engineconn-plugins/hive/pom.xml
@@ -29,7 +29,7 @@
     <artifactId>linkis-engineplugin-hive</artifactId>
 
     <properties>
-        <hive.version>1.2.1</hive.version>
+        <hive.version>3.1.2</hive.version>
     </properties>
 
     <dependencies>
